fix: Error when try cancel delete (#296)

+++ easyfinance.client/src/app/features/user/detail-user/detail-user.component.ts
@@ -121,8 +121,8 @@ export class DetailUserComponent implements OnInit {
     });
   }
 
-  confirmDeletion(): void {
-    if (this.deleteToken) {
+  confirmDeletion(result: boolean): void {
+    if (result && this.deleteToken) {
       this.userService.deleteUser(this.deleteToken).subscribe({
         next: (response) => {
           this.userService.removeUserInfo();
